Add multi-user support
We're assuming that
/nix/var/nix/profiles/default/etc/profile.d/nix-daemon.sh will only
exist on a multi-user install. It's possible to construct this path on a
single-user install but it would be exceedingly strange to do so.

This is safer than the inverse, of testing for
~/.nix-profile/etc/profile.d/nix.sh first, because in a multi-user
install there's nothing stopping the user from writing `nix-env -iA nix`
which will add that path to their profile.

Fixes #1.